requirements.txt
Pyrogram
TgCrypto
py-tgcalls==0.5.5
python-dotenv
youtube_dl
youtube_search_python
requests
aiohttp
aiofiles
asyncio
youtube_search
search_engine_parser
ffmpeg
Pillow
python-arq
future
psutil
wget
regex
dnspython
bs4
motor
tswift
python-telegram-bot==12.8
telegraph
telegram
